added test to compare each term -- all match except nonnegative barrier
